pull initialization effects out of ProjectView into its own component
