+++ src/views/project/ProjectView.tsx
@@ -1,38 +1,19 @@
-import { useEffect, useRef } from "react";
-import { useLocation } from "react-router-dom";
-import { projectRoute } from "../../lib/services/routeService";
+import { useCallback, useRef } from "react";
 import ProjectList from "./ProjectCards";
-import { useAppDispatch, useAppSelector } from "../../lib/hooks/reduxHooks";
-import {
-  projectGetShouldFetch,
-  projectInitialFetch,
-} from "../../store/projectSlice";
+
+import ProjectInitializer from "./ProjectInitializer";
 
 const ProjectView = () => {
-  const dispatch = useAppDispatch();
   const projectContainerRef = useRef<HTMLElement>(null);
-  const location = useLocation();
-  const shouldFetch = useAppSelector(projectGetShouldFetch);
-  /**
-   * Initial fetch
-   */
-  useEffect(() => {
-    if (shouldFetch) {
-      dispatch(projectInitialFetch());
-    }
-  }, [dispatch, shouldFetch]);
-  /**
-   * Scroll project view into view if pathname === '/project'
-   */
-  useEffect(() => {
-    if (location.pathname + location.hash === `${projectRoute.path}`) {
-      const projectContainer = projectContainerRef.current!;
-      projectContainer.scrollIntoView({ behavior: "smooth" });
-    }
-  }, [location]);
+
+  const scrollProjectUp = useCallback(() => {
+    const projectContainer = projectContainerRef.current!;
+    projectContainer.scrollIntoView({ behavior: "smooth" });
+  }, []);
 
   return (
     <section id="project" className="py-[7.5rem]" ref={projectContainerRef}>
+      <ProjectInitializer scrollProjectUp={scrollProjectUp} />
       <ProjectHeader />
       <ProjectList />
     </section>
